First pass at a fix for JENKINS-8592.
The problem is that the sending side has no way of knowing when the
receiving side has hung up. So when the receiving side hangs up we should
be able to read till the stream runs out, then close. Not the optimal
solution, but one that should work for now.

http://issues.jenkins-ci.org/browse/JENKINS-8592