Enabling user-level classloader support for properties.
Without this change, values added as properties must entirely consists of classes that are available on the remote peer, or else "getRemoteProperty" from the other side will fail.

What makes me bit nervous is that I vaguely recall that I intentionally disabled this during the implementation, and I can't exactly remember why.

I think it's partly because I was originally thinking of using it for capability exchange (or other similar lower-level services that the classloader forwarding relies on.) Such use would have created infinite recursion. But we aren't doing that yet. It's also perhaps partly for security, so that you can get properties without worrying about the remote peer injecting code, but that's now controlled via a separate switch, and it'll take effect here as well.

So I'm enabling this, along with several tests to go by.