[JENKINS-11251 JENKINS-9189] Resurrecting what's deleted in e0e154d12d7a10759287b187467389c6e643c12b
When communicating with remoting < 2.15, this allows them to continue to
perform some degree of syncing, so that they can still enjoy the fix for
JENKINS-9189.

None of these code is exposed via API outside remoting, so at some point
we can revert this change to simplify the code a bit and eliminate the
redundancy, because as long as >= 2.15 remoting talk to each other,
PipeWriter does everything we need.