Combine jar retrieval and class image fetching
The previous iteration of the classloaing improvement favored retrieval
of whole jar file over the individual class file image transfer such
that a jar retrieval was necessary even if a single class is being used.
This penalizes the first connection by requiring a large amount of data
download even for a simple task.

In this change, we restore the balance a bit by making the jar file
loading a background activity. IClassLoader.fetch3() sends the image
of the class file along with the jar file checksum, until the remote
side confirms that it possesses the said jar file.

The jar file downloading will be done using a single background thread
and the queue.

In this way, the activity that triggered a remote class loading can
proceed without getting blocked by the jar file retrieval.