[JENKINS-18405]
This fixes the 'Unknown url shema' error in the FindBugs plugin.

Before JENKINS-15120, we used to faithfully recreate the expected
resource path in a temporary resource file. For some reasons, we lost
that. This change brings it back by recreting the directory structure.

Note that this change doesn't address
"java.lang.IncompatibleClassChangeError: Implementing class" error
reported also in JENKINS-18405. That is still under investigation.