[FIXED JENKINS-18394]
As a part of the prefetching, Jenkins 1.519 parses classes and tries
to discover other classes that the current class relies on.
"IncompatibleClassChangeError: Implementing class" was encountered
during such analysis phase.

This is a real problem in FindBugs source tree, but the code in question
is never getting exercised, which is why it was working OK.

In this change, I made the prefetching analysis such that it will ignore
the failures encounted during the reference analysis.

I need to start testing plugins with cutting-edge Jenkins version to
detect errors like this before it hits the users.