[FIXED JENKINS-20769]
When ProxyWriter is finalized, it should attempt to release a remote reference count,
but it shouldn't try to close the stream. For one, the other side that had exported
the writer might be still using it (imagine exporting System.out), and secondly, the
said object could have already been unexported, for example if the closure that brought
ProxyWriter to this side has already returned, since RemoteWriter is auto-unexporting.

Compare ProxyOutputStream impl vs ProxyWriter impl and the difference becomes clearer.