Brought ProxyWriter on par with ProxyOutputStream.
Since ProxyOutputStream is used far more often than ProxyWriter, it looks like we've failed
to keep them both up to the same level.

This change ports newer features and fixes from ProxyOutputStream to ProxyWriter, including
flushing, throttling, and I/O synchronization.

Since the code looks so similar, one wonders if there is any way to unify them, but
this turns out to be tricky because of the need to remain backward compatible. Command objects
for example need to be available in the exact same name, which means if we are to try to
unify them, we'd end up creating 3 copies (one in ProxyOutputStream, onein ProxyWriter, and
one in the unified version.) So I'm sticking to two copies.