[JENKINS-20707]
I tried to write a test case to verify the hypothesis along the line of
https://issues.jenkins-ci.org/browse/JENKINS-20707?focusedCommentId=198755&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-198755,
but ultimately it was a dead-end.

The reason I was wrong is because ImportedClassLoaderTable maintain strong
references to all the imported classloaders, and it prevents them from
getting garbage collected. Probably for the reasons I was mentioning
in JENKINS-20707, RemoteClassLoaders never get garbage collected, so they should
never get unexported. I've spent much time trying to make the test case
work, so I'm going to commit it on the side just in case someone finds
useful in the future.