Introducing the mechanism to verify that the Callables we are getting is legitimate.
This change introduces Role, which designates the part one side plays in the overall protocol (such as "client" or "server".)
Callable and other Callable-like classes (like FileCallable seen in Jenkins) should implement this RoleSensitive interface
to declare the intended place the Callables run.

This information is then used as a basis for decision making to allow/refuse Callables. RoleChecker does this, and it gets
installed as a CallableDecorator.

This change also makes a bold move to have Callable implement the newly added RoleSensitive, breaking source compatibility.
Fortunately, this change doesn't affect binary compatibility, and the care is taken to warn the caller that it might see
AbstractMethodError.

This change forces each Callable implementation to think about where it can be run, which provides static check and helps
those users of this library that wants to enforce this role mechanism (such as Jenkins.) It is indeed a nuisance for users
who have symmetric protocol (for example, the most of unit tests in this library), but the trade-off is judged positive
as the most critical user of this library (Jenkins) has an asymmetric protocol.