ChunkedCommandTransport had broken the stdout evacuation.
Unlike ClassicCommandTransport, ChunkedCommandTransport wasn't
passing on the "underlying output stream", so the StandardOutputSwapper
class in Jenkins core wasn't able to evacuate fd=1.

This means if JVM writes to fd=1, for example because of a VM crash,
then we won't be able to capture that in the log output.
Instead it just corrupts the communication stream.

This whole "underlying output stream" is a hack. In the next commit
I'm going to use the channel property to fix this by using a property.
This fix is still needed because older versions of Jenkins expect the
property to return something.