[FIXED JENKINS-21251] Initial calculation of wait time was susseptable to clock drift.
The calculation on how long to wait in various places on the code was
suseptable to the clock changing as multiple calls where made to
System.currentTimeMillis() for the initial calculation.

Changed this so that for the initial calculation only one call will be
made to System.currentTimeMillis().
WHilst this does not guarantee that the actuall amount of time will have
been waited for (and subsequent calculations are suseptable) at least on
the hotstop VM Object.wait() will not have spurious wake-ups.
Therefore the overhead of moving to System.nanoTime() is currently not
considered to be worth the performance overhead.