Deflake JnlpProtocol2Test
PowerMockito wasn't being told to prep Properties.class to have its
'new Date()' call hijacked, so it wasn't actually doing it. That said,
even adding that didn't seem to help.

But instead of reaching into the implementation details, I've written a
custom Matcher to just verify that the output string representation of
the Properties parses back to the expected values. As an added benefit,
this no longer depends on two identical Properties objects writing keys
in the same order when store() is called, since the API there doesn't
actually guarantee that.