[SECURITY-218] Allow a user to specificy the blacklisted classes.
We want a user to be able to blacklist classes without unpacking wars or
changing code, so provide a mechanism for an administrator to set a
property (hudson.remoting.ClassFilter.DEFAULTS_OVERRIDE_LOCATION) to point
at a file that contains a list of regular expressions.  A match in the
RegExp signifies that the class should be blacklisted.