Fixed the signature change in ClassFilter.NOOP
I get that James doesn't like public static final fields, but in this
case I think the change is overzealous. Let's respect how this was
originally done and stick to it. There are multiple good reasons.

There's a principle of minimizing a change for security fixes.  This fix
is allowed to get into 1.625.3 through a slip-stream, so let's not make
changes that are not necessary.

Then there's added complexity in keeping the old signature working while
adding the new getDefaultFilter() method. The code as it was also
removed the final modifier from the field, which is not nice.

There's also no good reason why constants like NONE is ever going to be
something other than constants.

All in all I think the fix is not worth the complication at this point.