INFO logging interferes with console apps that uses remoting.
JVM puts INFO logging to stdout by default, so if it's used by a CLI app
that needs to precisely control its output, it'll mess that up.

This includes Jenkins CLI.

To make up for the loss of ability to inspect what the actual rules are,
add the toString() method, so that for example Jenkins users can run
"print ClassFilter.DEFAULT" to find out the actual rules, not just that
the custom properties are in use.