Better to fail instead of fall back.
If the user did specify an override file, it's better to fail hard than
fall back to the default in case of a problem.

The situation I'm afraid of is that when Jenkins is using this, you can
make an erranous change to the override file and fail to realize that
your configuration change is not taking effect. Noticing that requires
an active act of looking into log files, which people don't do.