Fixing findbugs warning.
This is a difficult call. but in the environment where the platform encoding is not UTF-8, users expect to write such files in the platform default encoding.

The downside is that the definition file is now non-portable, though non-ASCII class names are uncommon even among CJK users, and this file is really only going to be used in Jenkins masters, so portability is less important compared to something that needs to be on all slaves.