Need a line with known text to find failures
Server who doesn't understand the JNLP3 protocol sends us a line like
"Unknown protocol:....", whereas the protocol expects an integer.
This results in the following exception:

    SEVERE: For input string: "Unknown protocol:Protocol:JNLP3-connect"
    java.lang.NumberFormatException: For input string: "Unknown protocol:Protocol:JNLP3-connect"
            at java.lang.NumberFormatException.forInputString(NumberFormatException.java:65)
            at java.lang.Integer.parseInt(Integer.java:492)
            at java.lang.Integer.parseInt(Integer.java:527)
            at org.jenkinsci.remoting.engine.JnlpProtocol3.initiateAndValidateMaster(JnlpProtocol3.java:185)
            at org.jenkinsci.remoting.engine.JnlpProtocol3.performHandshake(JnlpProtocol3.java:141)
            at org.jenkinsci.remoting.engine.JnlpProtocol.establishChannel(JnlpProtocol.java:77)
            at hudson.remoting.Engine.run(Engine.java:308)

To gracefully handle this, we need JNLP3 protocol to send a known line
so that the client can easily distinguish a successful start of the
authentication/negotiation against a downright rejection. I'e added
the "Negotiate\n" line to do this here.

In JNLP1 and JNLP2, the first response thta the server sends is GREETING_SUCCESS,
so this problem doesn't apply. JNLP3 is the first protocol that requires
some back & forth, which revealed this hidden assumption.