Improve the caller/callee correlation diagnostics
I was trying to diagnose a remoting call that bridges two VMs that is
hanging. This can be made easier if thread dump allows us to correlate
the caller/callee.

This change adds the request ID to thread name so that thread dump will
show this clearly

Also fixes one incorrect use of toString() in diagnosis, which should
be getName() --- see all the other uses of thread names.