Unravel the classloading clew coming from the PreloadJarTask
It appears that the target classloader in PreloadJarTask may be actually null. And FindBugs is grumbling about it. So I made an attempt to process the dependencies and to properly annotate methods and to catch possible edge cases.

The only real problem is that UserRequest#perform() may fail horribly if the request has been constructed with null constructor. So I have added explicit handling for this case on the invoking side.