Remoting 2.x End Of Life
Remoting 2.x has reached its end of life on May 01, 2017. My plan is to release new version if required, but I want to oficially call it as a deprecated version in order to avoid proactive backporting and to simplify documentation.

What does change for Remoting 2.x users? If you need an issue backporting, they will have to request it