Fix [JENKINS-42533]
In rare cases, the Unexport command gets a little out of sequence and attempts to unexport an object that has already been unexported. It logs this at SEVERE level, when there is no indication of any harm. The issue is marked as minor.

Implement the fix in a similar pattern to [JENKINS-22853], as noted in the PR. That fix reduced the log level for a similar message coming from the EOF command. As with that fix, it is unclear how to create a meaningful test, particularly as it is only a log level change..