[JENKINS-53569] Remove unnecessary locking.
There is no indication that this section of code constitutes a critical section. The only meaningful thing it does is atomic and is handled elsewhere in the code by good design. The extra checks serve no significant purpose.
The code in Filter that calls this grabs a lock on Filter before it calls this meaning that it already holds one lock. If another thread grabs the lock on ProtocolStack before the Filter than a thread is always possible.
Probably more of the locks in this area should be cleaned up and the threading / locking model simplified.