Re-work the implementation of the no_proxy capability.
We've had several issues submitted where it hasn't provided the capabilities people need.
The most recent one is JENKINS-48778 but we fixed a very similar one a couple
of months ago. The existing implementation was a bit more difficult to follow
than necessary.
This one adds a few more capabilities, retains the existing ones, and
documents the supported formats.