Change new no_proxy impl to not draw in new dependencies.
For the changes to the no_proxy capability, I added a couple of dependencies for checking and handling ip addresses and subnets.
One of them was clearly the wrong one to use when I started integrating.
Rather, than add dependencies, we decided to copy in utilities. This same pattern was followed recently in Jenkins core to avoid introducing new dependencies into Jenkins core that plugins may become dependent on.