Update "@since TODO" tags.
I had a few that I needed to update for the 3.34 release.
In trying to do those, I noticed there are a bunch of old ones. There is very little value in the ones that are 2, 3, or certainly 5 years old. If I could quickly identify the correct value, I added it, otherwise I removed the tag.