make RemoteClassLoader more robust in case of remote interruption
Make RemoteClassLoader more robust in case of "RemotingSystemException caused by an InterruptedException".
This is done by adding a try/catch/retry dynamic proxy on top of the IClassLoader proxy. The logic itself
is similar to what was previously in place in some parts of RemoteClassLoader#findClass, but hopefully
this approach will cover more code paths (eg., interruption when loading a resource required for
initializing a class).