[JENKINS-67829] compute the checksum per JVM
as we do not support dynamic class unloading a URL to a checksum can
never change for the lifetime of the JVM.  That is in Jenkins we can
only install new plugins, upgrades always require a restart.

To that end we the Checksum cache in JarLoaderImpl always is the
checksum of the file on the local side so the only way it can change
is if the JVM is restarted.
Therefor we do not need to caclulate it for every channel, rather we can
create it once for all Channels on a given JVM.