[JENKINS-67829] compute the checksum per JVM
As access to CHECKSUMS and the computation of the Checksum was
non-atomic we could still have lots of threads attempting to checksum
the same jar file at startup.
Whilst the jar would likely end up in the disk cache and IO would not be
too high, the digest calculation would still occur causing wasted CPU
that could be used for other things.

Make the computation atomic wrt checking that we have the checksum